requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
PyYAML>=5.4.0
pytest>=7.0.0
pytest-mock>=3.0.0
